+++ framework/db-http/src/test/java/is/codion/framework/db/http/AbstractHttpEntityConnectionTest.java
@@ -386,6 +386,7 @@ abstract class AbstractHttpEntityConnectionTest {
 						.domainClasses(singletonList(TestDomain.class.getName()))
 						.sslEnabled(false)
 						.auxiliaryServerFactory(singletonList(EntityServiceFactory.class.getName()))
+						.objectInputFilterFactoryRequired(false)
 						.build();
 	}