+++ tools/monitor/ui/src/test/java/is/codion/tools/monitor/ui/EntityServerMonitorPanelTest.java
@@ -93,6 +93,7 @@ public class EntityServerMonitorPanelTest {
 						.adminUser(User.parse("scott:tiger"))
 						.connectionPoolUsers(Collections.singletonList(UNIT_TEST_USER))
 						.domainClasses(Collections.singletonList(TestDomain.class.getName()))
+						.objectInputFilterFactoryRequired(false)
 						.database(Database.instance())
 						.sslEnabled(false)
 						.build();